split module and create separate file for L-BFGS, mark TODOs for PR comments
